Move all_types_table
