+++ tests/conftest.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 import os
 import subprocess
-import textwrap
 import time
 
 import pytest
@@ -51,48 +50,6 @@ def mapd_server():
 def con(mapd_server):
     return connect(user="mapd", password='HyperInteractive', host='localhost',
                    port=6274, protocol='binary', dbname='mapd')
-
-
-@pytest.fixture
-def all_types_table(con):
-    """A table with all the supported types
-
-    - boolean_
-    - smallint_
-    - int_
-    - bigint_
-    - decimal_
-    - float_
-    - double_
-    - varchar_
-    - text_
-    - time_
-    - timestamp_
-    - date_
-
-    https://www.mapd.com/docs/latest/mapd-core-guide/tables/#create-table
-    """
-    name = 'all_types'
-    drop = 'drop table if exists {};'.format(name)
-    c = con.cursor()
-    c.execute(drop)
-    create = textwrap.dedent('''\
-    create table {name} (
-        boolean_ BOOLEAN,
-        smallint_ SMALLINT,
-        int_ INT,
-        bigint_ BIGINT,
-        float_ FLOAT,
-        double_ DOUBLE,
-        varchar_ VARCHAR(40),
-        text_ TEXT,
-        time_ TIME,
-        timestamp_ TIMESTAMP,
-        date_ DATE
-    );'''.format(name=name))
-    # skipping decimal for now
-    c.execute(create)
-    return name
 
 
 @pytest.fixture
